renamed labware folder to instruments
