+++ opentrons_sdk/instruments/__init__.py
@@ -1,0 +1,2 @@
+from opentrons_sdk.instruments.magbead import Magbead
+from opentrons_sdk.instruments.pipette import Pipette
