--- opentrons_sdk/labware/instruments.py
@@ -1,2 +1,0 @@
-from opentrons_sdk.labware.magbead import Magbead
-from opentrons_sdk.labware.pipette import Pipette
