+++ tests/opentrons_sdk/helpers/test_calibration.py
@@ -2,7 +2,7 @@ import json
 import os
 import unittest
 
-from opentrons_sdk.labware import instruments
+from opentrons_sdk import instruments
 from opentrons_sdk import containers
 from opentrons_sdk.robot import Robot
