+++ tests/opentrons_sdk/integration/test_protocol.py
@@ -3,7 +3,7 @@ from opentrons_sdk import containers
 
 from opentrons_sdk.robot import Robot
 from opentrons_sdk.containers.placeable import Container, Deck
-from opentrons_sdk.labware import instruments
+from opentrons_sdk import instruments
 
 
 class ProtocolTestCase(unittest.TestCase):
